+++ client-app/shared/layout/components/header/_internal/mobile-menu/menus/settings-menu.vue
